feat: wire check_alerts, fix alerts table, Polygon volume as f64
- Call check_alerts after successful fetch_ticker_data; poll quotes on Alerts tab
- draw_alerts: five Table constraints matching five columns (#37)
- TickerResult/HistoricalData volume as f64 for fractional Polygon JSON
- Docs: SPEC §7 shipment, ROADMAP gaps, QA scope; clear SCRATCHPAD for ship

Closes #30. Closes #37. Closes #38.

+++ src/models/historical.rs
@@ -20,7 +20,7 @@ pub struct HistoricalData {
     pub h: f64, // High
     pub l: f64, // Low
     pub c: f64, // Close
-    pub v: u64, // Volume
+    pub v: f64, // Volume (Polygon may return fractional values)
     pub t: u64, // Timestamp
     pub vw: f64, // Volume weighted average price
     pub n: Option<u64>, // Number of transactions
